+++ src/main/java/blackjack/model/bet/BetAmount.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.bet;
 
 public class BetAmount {
