+++ src/test/java/blackjack/model/BetAmountTest.java
@@ -2,6 +2,7 @@ package blackjack.model;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import blackjack.model.bet.BetAmount;
 import blackjack.model.user.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
